Force VS to use one single binary-dir
The standard-behaviour of MSVC to use 'Debug' and 'Release'-Folders
leads to several problems when your project requires you to move
data-files and/or .DLLs into these directories to make your project
run. The biggest problem is that during the CMake-Run this behaviour
makes it impossible to determine any target-directory for these files.

So this set of changes forces MSVC to do the same as every other com-
piler does: Using a single predefined output-directory for binaries.
Only drawback is: You need to do a clean-rebuild if switching from
Release- to Debug-Builds and vice versa.